upd1

+++ Kyrcoiva-Pamyatka/Kyrcoiva-Pamyatka.cpp
@@ -18,7 +18,7 @@ int main()
     cout << r << "ГРА ПАМ'ЯТЬ\n"<<o;
     Sleep(1000);
     cout << "Виберіть розмір поля "<<m<<b<< "4x4"<<mn<<o<<" - 1 або "<<m<<y<< "5x6"<<mn<<o<<" - 2:\n";
-    short choice;
+    int choice;
     cin >> choice;
     while (choice != 1 && choice != 2) {  
         cerr << "Виберіть " << r << "1" << o << " або " << r << "2" << o << " для початку гри!\n";
